refactor

+++ app/Http/Controllers/Admin/Product/ProductActivationKey/DestroyController.php
@@ -10,6 +10,8 @@ class DestroyController extends Controller
 {
     public function __invoke(Product $product, ActivationKey $activationKey)
     {
+        // policy
+        $this->authorize('delete', $activationKey);
         $activationKey->delete();
         return response()->json('Activation key removed', 200);
     }
